backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
google-genai
python-multipart>=0.0.6
requests>=2.31.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
reportlab>=4.0.0